# api_docs/api-docs-input.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /api/rooms/{roomId}/hotels/{hotelId}:
    put:
      tags:
      - room-rest
      summary: Edit a room
      description: Edit a room by room ID and hotel ID.
      operationId: editRoom
      parameters:
      - name: roomId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: hotelId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: updates
        in: query
        required: true
        schema:
          type: object
          additionalProperties:
            type: object
      responses:
        "404":
          description: Room or hotel not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Room_RoomDetails'
        "403":
          description: Forbidden operation
          content:
            application/json: {}
        "200":
          description: Room edited successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room_RoomDetails'
...
  /api/hotels/{id}:
    get:
      tags:
      - hotel-rest
      summary: Get hotel data
      description: Fetches the hotel data for the given ID if the current user is
        the manager of the hotel and the manager is validated.
      operationId: getHotelData
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "403":
          description: "Forbidden, current user is not the manager of the hotel"
          content:
            application/json: {}
        "200":
          description: Found the hotel
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hotel_HotelDetails'
        "404":
          description: Hotel not found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
    put:
      tags:
      - hotel-rest
      summary: Edit a hotel
      description: Edits the details of the hotel with the given ID if the current
        user is the manager of the hotel.
      operationId: editHotel
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: updates
        in: query
        required: true
        schema:
          type: object
          additionalProperties:
            type: object
      responses:
        "403":
          description: "Forbidden, current user is not the manager of the hotel"
          content:
            application/json: {}
        "200":
          description: Hotel edited successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hotel_HotelDetails'
        "404":
          description: Hotel not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Hotel_HotelDetails'
    delete:
      tags:
      - hotel-rest
      summary: Delete a hotel
      description: Deletes the hotel with the given ID if the current user is the
        manager of the hotel.
      operationId: deleteHotel
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "403":
          description: "Forbidden, current user is not the manager of the hotel"
          content:
            application/json: {}
        "204":
          description: Hotel deleted successfully
          content:
            application/json: {}
        "404":
          description: Hotel not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Hotel'
...
  /api/users/{id}/image:
    get:
      tags:
      - user-image-rest
      summary: Download user profile image
      operationId: downloadProfileImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User profile image downloaded successfully
          content:
            image/*: {}
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                type: object
    post:
      tags:
      - user-image-rest
      summary: Upload user profile image
      operationId: uploadImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              required:
              - imageFile
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
      responses:
        "403":
          description: Operation only allowed for the user or an admin
          content:
            application/json: {}
        "201":
          description: User profile image uploaded successfully
          content:
            application/json: {}
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - user-image-rest
      summary: Delete user profile image
      operationId: deleteImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "403":
          description: Operation only allowed for the user or an admin
          content:
            application/json: {}
        "204":
          description: User profile image deleted successfully
          content:
            application/json: {}
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                type: object
  /api/rooms/hotels/{hotelId}:
    post:
      tags:
      - room-rest
      summary: Create a room
      description: Create a new room for a specific hotel.
      operationId: postRoom
      parameters:
      - name: hotelId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Room'
        required: true
      responses:
        "201":
          description: Room created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        "403":
          description: Forbidden operation
          content:
            application/json: {}
        "404":
          description: Hotel not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Room'
  /api/reviews/hotels/{hotelId}:
    post:
      tags:
      - review-rest
      summary: Post a hotel review
      description: Post a review for a specific hotel by hotel ID.
      operationId: postReview
      parameters:
      - name: hotelId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'
        required: true
      responses:
        "400":
          description: Invalid input
          content:
            application/json: {}
        "201":
          description: Review created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        "404":
          description: Hotel not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Review'
...
  /api/hotels:
    get:
      tags:
      - hotel-rest
      summary: Load all hotels
      description: Returns a page of all hotels.
      operationId: loadAllHotels
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: Hotels retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_HotelDetails'
        "403":
          description: "Forbidden, the user does not have permission to view these\
            \ hotels"
          content:
            application/json: {}
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageResponseHotel_HotelDetails'
    post:
      tags:
      - hotel-rest
      summary: Create a hotel
      description: Creates a new hotel with the given details.
      operationId: createHotel
      parameters:
      - name: name
        in: query
        required: true
        schema:
          type: string
      - name: description
        in: query
        required: true
        schema:
          type: string
      - name: location
        in: query
        required: true
        schema:
          type: string
      - name: room1
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: cost1
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: room2
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: cost2
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: room3
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: cost3
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: room4
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: cost4
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Hotel created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hotel'
        "404":
          description: Manager not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Hotel'
  /api/hotels/{id}/image:
    get:
      tags:
      - hotel-image-rest
      summary: Download hotel image
      description: Download the image of a specific hotel by hotel ID.
      operationId: downloadProfileImage_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Hotel not found
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: Image retrieved successfully
          content:
            image/*: {}
    post:
      tags:
      - hotel-image-rest
      summary: Upload hotel image
      description: Upload an image for a specific hotel by hotel ID.
      operationId: uploadImage_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              required:
              - imageFile
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
      responses:
        "201":
          description: Image uploaded successfully
          content:
            application/json: {}
        "403":
          description: Forbidden operation
          content:
            application/json: {}
        "404":
          description: Hotel not found
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - hotel-image-rest
      summary: Delete hotel image
      description: Delete the image of a specific hotel by hotel ID.
      operationId: deleteImage_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Image deleted successfully
          content:
            application/json: {}
        "403":
          description: Forbidden operation
          content:
            application/json: {}
        "404":
          description: Hotel not found
          content:
            '*/*':
              schema:
                type: object
...
  /api/rooms/hotels/{id}:
    get:
      tags:
      - room-rest
      summary: Get hotel rooms
      description: Get all rooms of a hotel by hotel ID.
      operationId: hotelRooms
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: Rooms retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_RoomDetails'
        "403":
          description: Forbidden operation
          content:
            application/json: {}
        "404":
          description: Hotel not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageResponseRoom_RoomDetails'
...
  /api/reviews/size/hotels/{id}:
    get:
      tags:
      - review-rest
      summary: Get the number of hotel reviews
      description: Get the number of reviews of a specific hotel by hotel ID.
      operationId: hotelReviews
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "403":
          description: Forbidden operation
          content:
            application/json: {}
        "200":
          description: Number of reviews retrieved successfully
          content:
            application/json:
              schema:
                type: integer
                format: int32
        "404":
          description: Hotel not found
          content:
            '*/*':
              schema:
                type: integer
                format: int32
  /api/reviews/percentage/hotels/{id}:
    get:
      tags:
      - review-rest
      summary: Get the percentage of hotel reviews
      description: Get the percentage of reviews of a specific hotel by hotel ID.
      operationId: hotelReviews_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Percentage of reviews retrieved successfully
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden operation
          content:
            application/json: {}
        "404":
          description: Hotel not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: integer
                  format: int32
  /api/reviews/hotels/{id}:
    get:
      tags:
      - review-rest
      summary: Get hotel reviews
      description: Get all reviews of a specific hotel by hotel ID.
      operationId: hotelReviews_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: Reviews retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_ReviewDetails'
        "404":
          description: Hotel not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageResponseReview_ReviewDetails'
...
  /api/reservations/hotels/{id}:
    get:
      tags:
      - reservation-rest
      summary: Get hotel reservations
      description: Returns a page of reservations for a specific hotel.
      operationId: getHotelReservations
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: Reservations retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_ReservationDetails'
        "404":
          description: Hotel not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageResponseReservation_ReservationDetails'
...
components:
  schemas:
...
    Hotel_RoomDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        location:
          type: string
        getrating:
          type: number
          format: float
        image:
          type: boolean
        imagePath:
          type: string
        averageRating:
          type: number
          format: float
        numRooms:
          type: integer
          format: int32
        imageBase64:
          type: string
    Reservation_RoomDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        numPeople:
          type: integer
          format: int32
        rooms:
          $ref: '#/components/schemas/Room_RoomDetails'
        idReservation:
          type: integer
          format: int64
    Room_RoomDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        maxClients:
          type: integer
          format: int32
        getcost:
          type: number
          format: float
        reservations:
          type: array
          items:
            $ref: '#/components/schemas/Reservation_RoomDetails'
        hotel:
          $ref: '#/components/schemas/Hotel_RoomDetails'
...
    Review_ReviewDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        score:
          type: integer
          format: int32
        comment:
          type: string
        date:
          type: string
          format: date
        hotel:
          $ref: '#/components/schemas/Hotel_ReviewDetails'
        user:
          $ref: '#/components/schemas/UserE_ReviewDetails'
        idReview:
          type: integer
          format: int64
...
    Hotel_HotelDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        location:
          type: string
        getrating:
          type: number
          format: float
        image:
          type: boolean
        imagePath:
          type: string
        manager:
          $ref: '#/components/schemas/UserE_HotelDetails'
        rooms:
          type: array
          items:
            $ref: '#/components/schemas/Room_HotelDetails'
        reservations:
          type: array
          items:
            $ref: '#/components/schemas/Reservation_HotelDetails'
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review_HotelDetails'
        averageRating:
          type: number
          format: float
        numRooms:
          type: integer
          format: int32
        imageBase64:
          type: string
    Reservation_HotelDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        numPeople:
          type: integer
          format: int32
        rooms:
          $ref: '#/components/schemas/Room_HotelDetails'
        idReservation:
          type: integer
          format: int64
    Review_HotelDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        score:
          type: integer
          format: int32
        comment:
          type: string
        date:
          type: string
          format: date
        idReview:
          type: integer
          format: int64
    Room_HotelDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        maxClients:
          type: integer
          format: int32
        getcost:
          type: number
          format: float
    UserE_HotelDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        lastname:
          type: string
        bio:
          type: string
        location:
          type: string
        language:
          type: string
        phone:
          type: string
        email:
          type: string
        imagePath:
          type: string
        image:
          type: boolean
        organization:
          type: string
        getvalidated:
          type: boolean
        rejected:
          type: boolean
        collectionRols:
          type: array
          items:
            type: string
        rols:
          type: array
          items:
            type: string
        nick:
          type: string
        pass:
          type: string
        imageBase64:
          type: string
    Hotel:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        location:
          type: string
        getrating:
          type: number
          format: float
        image:
          type: boolean
        imagePath:
          type: string
        manager:
          $ref: '#/components/schemas/UserE'
        rooms:
          type: array
          items:
            $ref: '#/components/schemas/Room'
        reservations:
          type: array
          items:
            $ref: '#/components/schemas/Reservation'
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        averageRating:
          type: number
          format: float
        numRooms:
          type: integer
          format: int32
        imageBase64:
          type: string
    Reservation:
      type: object
      properties:
        id:
          type: integer
          format: int64
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        numPeople:
          type: integer
          format: int32
        hotel:
          $ref: '#/components/schemas/Hotel'
        room:
          $ref: '#/components/schemas/Room'
        user:
          $ref: '#/components/schemas/UserE'
        rooms:
          $ref: '#/components/schemas/Room'
        idReservation:
          type: integer
          format: int64
    Review:
      type: object
      properties:
        id:
          type: integer
          format: int64
        score:
          type: integer
          format: int32
        comment:
          type: string
        date:
          type: string
          format: date
        hotel:
          $ref: '#/components/schemas/Hotel'
        user:
          $ref: '#/components/schemas/UserE'
        idReview:
          type: integer
          format: int64
    Room:
      type: object
      properties:
        id:
          type: integer
          format: int64
        maxClients:
          type: integer
          format: int32
        getcost:
          type: number
          format: float
        reservations:
          type: array
          items:
            $ref: '#/components/schemas/Reservation'
        hotel:
          $ref: '#/components/schemas/Hotel'
    UserE:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        lastname:
          type: string
        bio:
          type: string
        location:
          type: string
        language:
          type: string
        phone:
          type: string
        email:
          type: string
        imagePath:
          type: string
        image:
          type: boolean
        organization:
          type: string
        getvalidated:
          type: boolean
        rejected:
          type: boolean
        collectionRols:
          type: array
          items:
            type: string
        rols:
          type: array
          items:
            type: string
        nick:
          type: string
        pass:
          type: string
        reservations:
          type: array
          items:
            $ref: '#/components/schemas/Reservation'
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        hotels:
          type: array
          items:
            $ref: '#/components/schemas/Hotel'
        imageBase64:
          type: string
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    PageResponse_HotelDetails:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    PageResponse_RoomDetails:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    PageResponseRoom_RoomDetails:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Room_RoomDetails'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    PageResponse_ReviewDetails:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    PageResponseReview_ReviewDetails:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Review_ReviewDetails'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    PageResponse_ReservationDetails:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    PageResponseReservation_ReservationDetails:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Reservation_ReservationDetails'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    PageResponseHotel_HotelDetails:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Hotel_HotelDetails'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32